Problematic fox files are now stored in the ./fox-error directory (overwrite with -x).
Better indication of where we're in the process with n/size added to the
log messages.

Also check if a resource exists with the import directory rewrite.

First generate the Dublin Core and use the first non-empty title as the
Fedora label, if that doesn't exist use the PID (MdSelfLink).