Update of FLAT to Fedora Commons 3.8.1, Islandora 7.x-1.6 modules, GSearch 2.8, SOLR 4.10.4.
Enables the use of the SOLR-based speedup of collection display.

Does enforce the move of CMD metadata to the compound DO. To get this to work the default object-view hook of the compound solution pack is disabled.

Switch from Saxon to Xalan for indexing, so FLAT can use the Discovery Garden templates.

Fix for the collapsible fieldset for datastreams in the object view. Now it works in a bootstrap theme (default of FLAT), but probably not in a default Drupal theme, e.g., Bartik.