Allow external location to contain ingestable data
So far, the deposit expected all data to be located in the drupal user
data directory. Now, this can also be an external location on the file
system on which drupal is running. The root to the user directory is
specified by the admin in the config.inc file, the name of the folder
within this user file is defined by the user during bundle initiation.